requirements.txt
# 基础工具库
numpy==1.26.4
requests==2.31.0
python-dotenv==1.1.1
typing-inspection==0.4.1

# 向量数据库
faiss-cpu==1.11.0

# 文档处理
PyPDF2==3.0.1
pdfplumber==0.10.2
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.1
html2text==2020.1.16
beautifulsoup4==4.12.2

# 嵌入模型
tenacity==8.2.3
torch==2.2.0
transformers==4.35.2
openai==1.3.5
zhipuai==2.1.0
dashscope==1.14.0

# LLM相关
pydantic==2.7.0
pydantic-settings==2.10.1
tiktoken==0.5.2

# 其他工具
tqdm==4.66.1